back/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
pydantic[email]==2.9.2
bcrypt==4.2.0
PyJWT==2.9.0
minio==7.2.9
python-multipart==0.0.12

# Features
redis==5.0.1
slowapi==0.1.9
weasyprint==62.1
python-docx==1.1.0
python-pptx

# WebSocket server (process séparé)
httpx==0.27.2
websockets==13.1
